+++ lib/anywhere_loader.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
+import 'dart:io';
 import 'dart:ui';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 /// A globally accessible loader that can be shown anywhere in the app
@@ -26,6 +28,12 @@ class AnyWhereLoader {
   double? _fontSize;
   Color? _fontColor;
   String? _fontFamily;
+  Color _progressColor = Colors.blue;
+  Color _backgroundColor = Colors.black;
+  double _opacity = 0.3;
+  double _blurSigma = 5.0;
+  bool _isIos=false;
+  double _iosProgressRadius=13;
 
   /// Starts the loader with optional customization and timeout
   void startLoader({
@@ -35,6 +43,12 @@ class AnyWhereLoader {
     Color? fontColor,
     String? fontFamily,
     Widget? customWidget,
+    Color? progressColor,
+    Color? backgroundColor,
+    double? opacity,
+    double? blurSigma,
+    bool? isIos,
+    double? iosProgressRadius
   }) {
     if (_isShowing) return; // Prevent duplicate loaders
 
@@ -44,7 +58,12 @@ class AnyWhereLoader {
     _fontSize = fontSize;
     _fontColor = fontColor;
     _fontFamily = fontFamily;
-
+    _progressColor = progressColor ?? Colors.blue;
+    _backgroundColor = backgroundColor ?? Colors.black;
+    _opacity = opacity ?? 0.3;
+    _blurSigma = blurSigma ?? 5.0;
+    _isIos=isIos??false;
+    _iosProgressRadius=iosProgressRadius??13;
     // Create and show the overlay
     _overlayEntry = _createOverlay();
     Overlay.of(context!).insert(_overlayEntry!);
@@ -73,6 +92,10 @@ class AnyWhereLoader {
     Color? fontColor,
     String? fontFamily,
     Widget? customWidget,
+    Color? progressColor,
+    Color? backgroundColor,
+    double? opacity,
+    double? blurSigma,
   }) async {
     startLoader(
       text: text,
@@ -81,6 +104,10 @@ class AnyWhereLoader {
       fontColor: fontColor,
       fontFamily: fontFamily,
       customWidget: customWidget,
+      progressColor: progressColor,
+      backgroundColor: backgroundColor,
+      opacity: opacity,
+      blurSigma: blurSigma,
     );
     try {
       await asyncFunction();
@@ -94,16 +121,13 @@ class AnyWhereLoader {
     return OverlayEntry(
       builder: (context) => Positioned.fill(
         child: BackdropFilter(
-          filter: ImageFilter.blur(sigmaX: 5, sigmaY: 5), // Blur effect
+          filter: ImageFilter.blur(sigmaX: _blurSigma, sigmaY: _blurSigma), // Blur effect
           child: Container(
-            color: Color.fromRGBO(0, 0, 0, 0.3), // Semi-transparent background
+            color: _backgroundColor.withAlpha((_opacity.clamp(0.0, 1.0) * 255).toInt()), // Custom opacity
             child: Center(
               child: Material(
                 color: Colors.transparent,
-                child: Padding(
-                  padding: const EdgeInsets.all(20),
-                  child: _customWidget ?? _defaultLoader(),
-                ),
+                child: _customWidget ?? _defaultLoader(),
               ),
             ),
           ),
@@ -114,20 +138,30 @@ class AnyWhereLoader {
 
   /// Default loading indicator with spinner and text
   Widget _defaultLoader() {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        const CircularProgressIndicator(),
-        const SizedBox(height: 10),
-        Text(
-          _text ?? "",
-          style: TextStyle(
-            fontSize: _fontSize ?? 16,
-            color: _fontColor ?? Colors.white,
-            fontFamily: _fontFamily,
+    return Padding(
+      padding: const EdgeInsets.all(20),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          _isIos?
+          CupertinoActivityIndicator(
+            radius: _iosProgressRadius,
+            color: _progressColor,
+          )
+              :CircularProgressIndicator(
+            valueColor: AlwaysStoppedAnimation<Color>(_progressColor),
           ),
-        ),
-      ],
+          const SizedBox(height: 10),
+          Text(
+            _text ?? "",
+            style: TextStyle(
+              fontSize: _fontSize ?? 16,
+              color: _fontColor ?? Colors.white,
+              fontFamily: _fontFamily,
+            ),
+          ),
+        ],
+      ),
     );
   }
 }